_old_code/apps/ai-service/requirements.txt
fastapi
uvicorn
python-multipart
gradio
ultralytics
langchain-core
langchain-ollama
Pillow
langchain-huggingface
sentence-transformers
chromadb
langchain_community
python-dotenv
pika
timm